Updates various tool functions so reports with multi-block Bases work.
"work" is generous here - this commit makes it possible to create
a HTML report using a model whose .basis has multiple tensor-product-
blocks (e.g. for leakage modeling).  It still doesn't work with
everything, in particular we punt on a couple reportables.py
functions where we just return -1 if the basis has >1 blocks.

+++ packages/pygsti/report/workspaceplots.py
@@ -2229,7 +2229,7 @@ class ProjectionsBoxPlot(WorkspacePlot):
         if M is None: M =  absMax
     
         d2 = len(projections) # number of projections == dim of gate
-        d = int(_np.sqrt(d2)) # dim of density matrix
+        d = _np.sqrt(d2) # dim of density matrix
         nQubits = _np.log2(d)
     
         if not _np.isclose(round(nQubits),nQubits):
